requirements.txt
altair==5.5.0
altgraph==0.17.4
attrs==25.3.0
blinker==1.9.0
bottle==0.13.4
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
macholib==1.16.3
MarkupSafe==3.0.2
narwhals==2.0.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
proxy_tools==0.1.0
pyarrow==21.0.0
pydeck==0.9.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
python-dateutil==2.9.0.post0
pytz==2025.2
pywebview==5.4
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.48.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0

# macOS-only: required by pywebview Cocoa/WebKit backend
pyobjc; platform_system == "Darwin"
